automation/requirements.txt
# Google Sheets API
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# YouTube API
google-api-python-client==2.108.0

# 既存の依存関係（親ディレクトリのrequirements.txtから）
requests==2.31.0
Pillow==10.1.0

# ログとユーティリティ
python-dotenv==1.0.0